+++ adapters/Nend/NendAdapter/GADMAdapterNendConstants.h
@@ -1,4 +1,4 @@
-// Copyright 2016 Google Inc.
+// Copyright 2019 Google Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,10 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import GoogleMobileAds;
+// Adapter version string
+static NSString *const kGADMAdapterNendVersion = @"5.1.0.0";
 
-/// Adapter for communicating with Facebook's Audience Network to fetch banner, interstitial,
-/// reward-based video, and native ads.
-@interface GADMAdapterFacebook : NSObject <GADMAdNetworkAdapter>
+// Adapter mediation agent
+static NSString *const kGADMAdapterNendMediationName = @"AdMob";
 
-@end
+// Nend API key
+static NSString *const kGADMAdapterNendApiKey = @"apiKey";
+
+// Nend spot id name
+static NSString *const kGADMAdapterNendSpotID = @"spotId";
+
+static NSString *const kGADMAdapterNendErrorDomain = @"com.google.mediation.nend";